Refactor of the questionnaires/rest_framework/views/public.py
