+++ api/app/signals/apps/questionnaires/rest_framework/views/__init__.py
@@ -1,13 +1,2 @@
 # SPDX-License-Identifier: MPL-2.0
 # Copyright (C) 2021 Gemeente Amsterdam
-from signals.apps.questionnaires.rest_framework.views.public import (
-    PublicQuestionnaireViewSet,
-    PublicQuestionViewSet,
-    PublicSessionViewSet
-)
-
-__all__ = (
-    'PublicQuestionnaireViewSet',
-    'PublicQuestionViewSet',
-    'PublicSessionViewSet',
-)
