+++ api/app/signals/apps/questionnaires/rest_framework/views/public/__init__.py
@@ -1,0 +1,13 @@
+# SPDX-License-Identifier: MPL-2.0
+# Copyright (C) 2021 Gemeente Amsterdam
+from signals.apps.questionnaires.rest_framework.views.public.questionnaires import (
+    PublicQuestionnaireViewSet
+)
+from signals.apps.questionnaires.rest_framework.views.public.questions import PublicQuestionViewSet
+from signals.apps.questionnaires.rest_framework.views.public.sessions import PublicSessionViewSet
+
+__all__ = [
+    'PublicSessionViewSet',
+    'PublicQuestionnaireViewSet',
+    'PublicQuestionViewSet',
+]
